confirmação

+++ php/posts.php
@@ -19,7 +19,7 @@
                             <a href='editar.php?id=$row[id]'><i class='fa-solid fa-pen-to-square'></i></a>
                         </div>
                         <div class='lixo'>
-                            <a href='../sql/delete.php?id=$row[id]'><i  class='fa-solid fa-trash-can'></i></a>
+                            <a href='../sql/delete.php?id=$row[id]' onclick=\"return confirm('Tem certeza que deseja deletar?');\"><i  class='fa-solid fa-trash-can'></i></a>
                         </div>
                     </div>";
                 }
